+++ NEWS.md
@@ -1,5 +1,7 @@
 # ggstats (development version)
 
+* Bug fix in `geom_stripped_rows()` and `geom_stripped_cols()` (#20)
+
 # ggstats 0.2.0
 
 * Support for pairwise contrasts (#14)